refactor: merge governance and scan assessment in one script
