--- output/.~lock.heatmap.csv#
@@ -1,1 +1,0 @@
-,straylite,orion,05.09.2025 19:30,file:///home/straylite/.config/libreoffice/4;
