--- output/.~lock.mapped-findings.csv#
@@ -1,1 +1,0 @@
-,straylite,orion,05.09.2025 15:50,file:///home/straylite/.config/libreoffice/4;
