+++ output/governance_heatmap_csv.csv
@@ -1,0 +1,14 @@
+csf_subcategory_id,name,response,severity,weighted_score
+ID.SC-01,Supply chain agreements are managed,No,high,1.00
+RC.GV-01,Recovery policies and procedures are established,No,high,1.00
+RC.IM-01,Recovery strategies are improved following exercises,No,high,1.00
+RC.IM-02,Recovery plans incorporate lessons learned,No,high,1.00
+RC.RP-01,Recovery plan is executed,No,high,1.00
+RS.CO-02,Public relations are managed,No,high,1.00
+RS.CO-03,Legal and regulatory requirements are managed,No,high,1.00
+RS.IM-01,Response plans are improved following lessons learned,No,high,1.00
+RS.MI-01,Incidents are mitigated,No,high,1.00
+RS.RP-01,Response plan is executed,No,high,1.00
+RC.CO-01,Communications are coordinated during recovery,No,high,0.80
+ID.GV-01,"Governance roles, policies, and processes are established",Partial,medium,0.50
+ID.AM-05,Resources are prioritized and protected,Yes,low,0.00
