+++ tests/test_nuc2csf.py
@@ -11,7 +11,7 @@ sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
 
 def test_get_paths():
-    from tools.nuc2csf import get_paths
+    from tools.assess import get_paths
 
     expected_paths = {  # Input Data
         #
@@ -50,7 +50,7 @@ def test_read_scan_json_valid():
 
     # Mock the open function and patch it
     with patch("builtins.open", mock_open(read_data=mock_json_content)):
-        from tools.nuc2csf import read_scan_json
+        from tools.assess import read_scan_json
 
         # Call the function and assert the result
         result = read_scan_json("mock_file.json")
@@ -63,7 +63,7 @@ def test_read_scan_json_invalid():
 
     # Mock the open function and patch it
     with patch("builtins.open", mock_open(read_data=mock_json_content)):
-        from tools.nuc2csf import read_scan_json
+        from tools.assess import read_scan_json
 
         # Assert that a JSONDecodeError is raised
         try:
@@ -89,7 +89,7 @@ TEMPLATE-2,Protect,PR.AC-1,Access Control
 
 def test_map_scan_to_csf(mock_lookup_csv):
     # import necessary functions
-    from tools.nuc2csf import map_scan_to_csf
+    from tools.assess import map_scan_to_csf
 
     # Sample findings (must be list of dicts, not strings)
     test_findings = [
@@ -121,7 +121,7 @@ def test_map_scan_to_csf(mock_lookup_csv):
 
 def test_write_with_data(tmp_path):
     # import necessary functions
-    from tools.nuc2csf import write_to_csv
+    from tools.assess import write_to_csv
 
     test_data = [{"name": "Alice", "age": 30}, {"name": "Bob", "age": 25}]
     path = tmp_path / "test.csv"
@@ -140,7 +140,7 @@ def test_write_with_data(tmp_path):
 
 def test_write_no_data(tmp_path):
     # import necessary functions
-    from tools.nuc2csf import write_to_csv
+    from tools.assess import write_to_csv
 
     path = tmp_path / "empty.csv"
     status = write_to_csv([], str(path))
@@ -150,7 +150,7 @@ def test_write_no_data(tmp_path):
 
 def test_write_to_json_with_data(tmp_path: Path):
     # Import necessary functions
-    from tools.nuc2csf import write_to_json
+    from tools.assess import write_to_json
 
     # prep test data and file
     data = [{"a": 1, "b": 2}]
@@ -168,7 +168,7 @@ def test_write_to_json_with_data(tmp_path: Path):
 
 def test_write_to_json_no_data(tmp_path: Path):
     # import necessary functions
-    from tools.nuc2csf import write_to_json
+    from tools.assess import write_to_json
 
     # prep test data and file
     data = []
@@ -183,7 +183,7 @@ def test_write_to_json_no_data(tmp_path: Path):
 
 
 def test_generate_scan_heatmap_handles_info_and_critical(tmp_path: Path):
-    from tools.nuc2csf import generate_scan_heatmap
+    from tools.assess import generate_scan_heatmap
     import numpy as np
 
     mapped = [
@@ -208,7 +208,7 @@ def test_generate_scan_heatmap_handles_info_and_critical(tmp_path: Path):
 
 
 def test_get_csf_lookup(tmp_path: Path):
-    from tools.nuc2csf import get_csf_lookup
+    from tools.assess import get_csf_lookup
 
     csv_content = (
         "csf_subcategory_id,weight,recommendation\n"
@@ -225,14 +225,14 @@ def test_get_csf_lookup(tmp_path: Path):
 
 
 def test_get_csf_lookup_missing_file():
-    from tools.nuc2csf import get_csf_lookup
+    from tools.assess import get_csf_lookup
 
     with pytest.raises(FileNotFoundError):
         get_csf_lookup("missing.csv")
 
 
 def test_get_governance_checklist_results(tmp_path: Path):
-    from tools.nuc2csf import get_governance_checklist_results
+    from tools.assess import get_governance_checklist_results
 
     checklist_content = (
         "csf_function,csf_subcategory_id,csf_subcategory_name,notes,response\n"
@@ -248,14 +248,14 @@ def test_get_governance_checklist_results(tmp_path: Path):
 
 
 def test_get_governance_checklist_results_missing_file():
-    from tools.nuc2csf import get_governance_checklist_results
+    from tools.assess import get_governance_checklist_results
 
     with pytest.raises(FileNotFoundError):
         get_governance_checklist_results("no-file.csv")
 
 
 def test_get_governance_checklist_results_missing_columns(tmp_path: Path):
-    from tools.nuc2csf import get_governance_checklist_results
+    from tools.assess import get_governance_checklist_results
 
     checklist_content = "csf_function,csf_subcategory_id,response\nIdentify,ID.AM-1,yes\n"
     checklist_path = tmp_path / "checklist.csv"
@@ -266,7 +266,7 @@ def test_get_governance_checklist_results_missing_columns(tmp_path: Path):
 
 
 def test_generate_governance_assessment():
-    from tools.nuc2csf import generate_governance_assessement
+    from tools.assess import generate_governance_assessement
 
     checklist = [
         {
@@ -306,7 +306,7 @@ def test_generate_governance_assessment():
 
 
 def test_generate_governance_heatmap():
-    from tools.nuc2csf import generate_governance_heatmap
+    from tools.assess import generate_governance_heatmap
 
     assessment = [
         {
